Medium
here is a collection of input strings and a collection of query strings. For each query string, determine how many times it occurs in the list of input strings. Return an array of the results.

Example



There are  instances of ',  of '' and  of ''. For each query, add an element to the return array, .
Link: https://www.hackerrank.com/challenges/sparse-arrays/problem?h_r=next-challenge&h_v=zen